requirements.txt
typing-extensions
langchain-core
langchain-openai
langchain-experimental
pandas
yfinance
stockstats
langgraph
rank-bm25
setuptools
backtrader
parsel
requests
tqdm
pytz
redis
chainlit
rich
typer
questionary
langchain_anthropic
langchain-google-genai
pydantic
python-dotenv
